refactor: ChessGame 패키지 수정
