+++ src/test/java/softeer2nd/chess/ChessGameTest.java
@@ -3,7 +3,6 @@ package softeer2nd.chess;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import softeer2nd.chess.pieces.ChessGame;
 import softeer2nd.chess.pieces.Piece;
 
 import java.util.List;
